# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SalonSeal API - No-show deposit system for Nairobi salons
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salons/{salonId}/waitlist:
    get:
      operationId: getWaitlist
      tags: [waitlist]
      summary: List all waitlist entries for a salon
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Waitlist entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistListResponse"
    post:
      operationId: addToWaitlist
      tags: [waitlist]
      summary: Add a client to the waitlist for a specific slot
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToWaitlistBody"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
  /salons/{salonId}/waitlist/{id}:
    delete:
      operationId: removeFromWaitlist
      tags: [waitlist]
      summary: Remove a waitlist entry (dismiss after notifying)
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /salons/{salonId}/waitlist/{id}/notify:
    patch:
      operationId: markWaitlistNotified
      tags: [waitlist]
      summary: Mark a waitlist entry as notified
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"

  # ── Reminders ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        staffId:
          type: ["integer", "null"]
        serviceId:
          type: ["integer", "null"]
        appointmentAt:
          type: string
          format: date-time
        clientName:
          type: string
        clientPhone:
          type: string
        serviceName:
          type: string
        staffName:
          type: ["string", "null"]
        notified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - appointmentAt
        - clientName
        - clientPhone
        - serviceName
        - notified
        - createdAt
    WaitlistListResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/WaitlistEntry"
      required:
        - entries
    AddToWaitlistBody:
      type: object
      properties:
        staffId:
          type: integer
        serviceId:
          type: integer
        appointmentAt:
          type: string
          format: date-time
        clientName:
          type: string
        clientPhone:
          type: string
        serviceName:
          type: string
        staffName:
          type: string
      required:
        - serviceId
        - appointmentAt
        - clientName
        - clientPhone
        - serviceName
...